Rename CheckedBox -> CheckBox
