+++ ui/DesignSystem/Primitive/Icon/CheckBox.svelte
